Delete bid impl early return when called even before the any bid was posted

+++ src/d3a/models/strategy/__init__.py
@@ -432,6 +432,8 @@ class BidEnabledStrategy(BaseStrategy):
             deleted_bid_ids = [bid.id for bid in self.get_posted_bids(market)]
         else:
             deleted_bid_ids = [bid_id]
+        if len(deleted_bid_ids) == 0:
+            return
         for b_id in deleted_bid_ids:
             if b_id in market.bids.keys():
                 market.delete_bid(b_id)
